requirements.txt
BeautifulSoup==3.2.1
CacheControl==0.11.7
Jinja2==2.6
PyYAML==3.11
Pygments==1.4
Sphinx==1.1.2
Unidecode==0.04.19
argparse==1.4.0
cement==2.10.2
coverage==3.7.1
docutils==0.8.1
elex==2.3.1
latimes-calculate==0.1.8
lockfile==0.12.2
ordereddict==1.1
pep8==1.7.0
pyflakes==1.3.0
pymongo==3.3.0
python-coveralls==2.4.2
python-dateutil==2.5.3
python-elections==1.1.3
python-slugify==1.2.1
requests==2.11.1
sh==1.09
six==1.6.1
ujson==1.35
wsgiref==0.1.2
